requirements.txt
streamlit
PyPDF2
python-dotenv

openai==0.28.1

langchain==0.0.354
chromadb==0.4.15
